feat-fix: decouple scopes in dataflow dependencies
