+++ src/dataflow/environments/index.ts
@@ -4,3 +4,4 @@ export * from './scoping'
 export * from './overwrite'
 export * from './environment'
 export * from './register'
+export { DataflowScopeName } from './scopes'
