+++ src/dataflow/environments/register.ts
@@ -1,6 +1,5 @@
 import { cloneEnvironments, IdentifierDefinition, IEnvironment, REnvironmentInformation } from './environment'
-import { DataflowScopeName } from '../graph'
-import { GlobalScope, LocalScope } from './scoping'
+import { DataflowScopeName, GlobalScope, LocalScope } from './scopes'
 
 /**
  * Insert the given `definition` --- defined within the given scope --- into the passed along `environments` will take care of propagation.
