+++ test/dataflow/elements/loops/for-loop.ts
@@ -1,6 +1,7 @@
 import { assertDataflow, withShell } from '../../../helper/shell'
 import { DataflowGraph, EdgeType, initializeCleanEnvironments } from '../../../../src/dataflow'
-import { appendEnvironments, define, LocalScope } from '../../../../src/dataflow/environments'
+import { appendEnvironments, define } from '../../../../src/dataflow/environments'
+import { LocalScope } from '../../../../src/dataflow/environments/scopes'
 
 describe('for', withShell(shell => {
 	assertDataflow(`Single-vector for Loop`,
