+++ test/dataflow/environments/initialization.ts
@@ -1,6 +1,6 @@
 import { DefaultEnvironmentMemory, Environment, initializeCleanEnvironments } from '../../../src/dataflow'
 import { expect } from 'chai'
-import { GlobalScope } from '../../../src/dataflow/environments'
+import { GlobalScope } from '../../../src/dataflow/environments/scopes'
 
 describe('Initialization', () => {
 	it('Clean creation should have no info but the default information', () => {
